# core/src/main/resources/swagger.yaml
...
swagger: '2.0'
info:
  title: Apache MetaModel Membrane
  description: Data Federation as a RESTful service.
  version: "1.0"
  contact:
    name: Apache MetaModel
    url: http://metamodel.apache.org
  license:
    name: Apache License, version 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0
consumes:
  - application/json
produces:
  - application/json
paths:
...
  /{tenant}/{datasource}/s/{schema}/t/{table}:
    parameters:
      - name: tenant
        in: path
        type: string
        description: The tenant name
        required: true
      - name: datasource
        in: path
        type: string
        description: The datasource name
        required: true
      - name: schema
        in: path
        type: string
        description: The schema name
        required: true
      - name: table
        in: path
        type: string
        description: The table name
        required: true
    get:
      responses:
        200:
          description: Table found
          schema:
            $ref: "#/definitions/getTableResponse"
        404:
          description: Table not found
          schema:
            $ref: "#/definitions/error"
  /{tenant}/{datasource}/s/{schema}/t/{table}/d:
    parameters:
      - name: tenant
        in: path
        type: string
        description: The tenant name
        required: true
      - name: datasource
        in: path
        type: string
        description: The datasource name
        required: true
      - name: schema
        in: path
        type: string
        description: The schema name
        required: true
      - name: table
        in: path
        type: string
        description: The table name
        required: true
    get:
      description: Gets the data of the table
      responses:
        200:
          description: Query executed
          schema:
            $ref: "#/definitions/queryResponse"
        400:
          description: Failure while parsing query
          schema:
            $ref: "#/definitions/error"
        404:
          description: Table not found
          schema:
            $ref: "#/definitions/error"
        500:
          description: Failure while executing query
          schema:
            $ref: "#/definitions/error"
    post:
      description: Inserts data to the table
      parameters:
        - name: inputData
          in: body
          description: The data to insert
          required: true
          schema:
            $ref: "#/definitions/insertionRequest"
      responses:
        200:
          description: Data inserted
          schema:
            $ref: "#/definitions/insertionResponse"
        404:
          description: Table not found
          schema:
            $ref: "#/definitions/error"
  /{tenant}/{datasource}/s/{schema}/t/{table}/c/{column}:
    parameters:
      - name: tenant
        in: path
        type: string
        description: The tenant name
        required: true
      - name: datasource
        in: path
        type: string
        description: The datasource name
        required: true
      - name: schema
        in: path
        type: string
        description: The schema name
        required: true
      - name: table
        in: path
        type: string
        description: The table name
        required: true
      - name: column
        in: path
        type: string
        description: The column name
        required: true
    get:
      description: Gets information about a column
      responses:
        200:
          description: Query executed
          schema:
            $ref: "#/definitions/getColumnResponse"
        404:
          description: Column not found
          schema:
            $ref: "#/definitions/error"
definitions:
  queryResponse:
    description: Represents the result of a query - a dataset
    type: object
    properties:
      type:
        type: string
        description: The type of entity (dataset)
      headers:
        type: array
        description: The dataset header names
        items:
          type: string
      data:
        type: array
        description: The actual data returned by the query
        items:
          type: array
          items:
            type: object
...
  insertionRequest:
    type: array
    items:
      description: A record to insert where each key is expected to match a column name and each value is the value to put.
      type: object
  insertionResponse:
    description: Represents the result of inserting records to a table
    type: object
    properties:
      status:
        type: string
        description: A confirmation 'ok' that the insertion went well.
      inserted-rows:
        type: number
        format: int32
        description: The amount of inserted records
      generated-keys:
        type: array
        description: An array of generated keys for the records, if any
        items:
          type: object
...
  getTableResponse:
    type: object
    properties:
      type:
        type: string
        description: The type of entity (table)
      name:
        type: string
        description: The table name
      schema:
        type: string
        description: The schema name
      datasource:
        type: string
        description: The datasource name
      tenant:
        type: string
        description: The tenant name
      columns:
        type: array
        description: The names of the table's columns
        items:
          type: object
          properties:
            name:
              type: string
              description: The column name
            uri:
              type: string
              description: A link to the column information
              format: uri
  getColumnResponse:
    type: object
    properties:
      type:
        type: string
        description: The type of entity (column)
      name:
        type: string
        description: The column name
      table:
        type: string
        description: The table name
      schema:
        type: string
        description: The schema name
      datasource:
        type: string
        description: The datasource name
      tenant:
        type: string
        description: The tenant name
      metadata:
        type: object
        description: Metadata about the column
        properties:
          number:
            type: integer
            description: The column number (0-based)
          size:
            type: integer
            description: The column size
          nullable:
            type: boolean
            description: Is the column nullable?
          primary-key:
            type: boolean
            description: Is the column a primary key?
          indexed:
            type: boolean
            description: Is the column indexed?
          column-type:
            type: string
            description: The column type (as interpreted/adapted by Apache MetaModel)
          native-type:
            type: string
            description: The native column type (as defined by the datasource itself)
          remarks:
            type: string
            description: Any remarks on the column
  error:
    description: Elaborates the error that occurred
    type: object
    properties:
      code:
        type: integer
        description: The HTTP status code
      message:
        type: string
        description: A humanly readable error message
      additional_details:
        type: object
        description: Any auxilary details to further elaborate the error